Add spring-integration-feed & spring-integration-file dependencies
The Spring Integration getting started suggests to:

    Download and unzip the source repository for this guide, or clone it using Git: git clone https://github.com/spring-guides/gs-integration.git
    cd into `gs-integration/initial`
    Jump ahead to `Define an Integration Flow`.

This path skip the Add to the Build Files section, which ask to add spring-integration-feed & spring-integration-file dependencies.

Therefore, someone following the guide might encounter an error. Based on 14, it looks it is expected to have already these dependencies declared in the build files.