Add dependabot configuration file
This commit adds dependabot configuration for keeping
versions up to date. To enable this, gradle build files
were also changed so that the dependency-management
plugin version is now managed by Boot and not set
explicitly. This commit also uses a common GitHub
Action and removes files used for Jenkins builds that
are no longer operational.